moved get & generate random float funcs to separate files
